64bit fixes + fedora build fix
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@1949 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ include/StringUtils.h
@@ -138,7 +138,7 @@ std::string		GetNextWord(std::string::const_iterator it, const std::string& str)
 bool 			Compress( const std::string & in, std::string * out, bool noCompression = false );	// Compress given string using zlib, noCompression will just add zlib header and checksum
 bool 			Decompress( const std::string & in, std::string * out );	// Decompress, returns false if checksum fails
 size_t			StringChecksum( const std::string & data );
-bool			FileChecksum( const std::string & path, uint * _checksum, size_t * _filesize );
+bool			FileChecksum( const std::string & path, size_t * _checksum, size_t * _filesize );
 
 
 ////////////////////
